Added basic combat, improved random room generation
Monsters can be attacked and killed. There are some bugs but they are
difficult to pin down.
- After killing a monster, some monsters cannot be attacked.
- The room Description does not currently update to reflect the changed
number of monsters in it.

Actual combat mechanics are sequestered into the GameRules static class.
This class will contain static methods to provide the logic for all
major game actions. Changing the rule set or game mechanics will rely
mostly on editing this class.